Abstracting the contract with the resources for AlphaVantage
