+++ systematic-trading-data-api-alpha-vantage/src/main/java/com/systematic/trading/signals/data/api/alpha/vantage/dao/impl/HttpAlphaVantageApiDao.java
@@ -46,15 +46,17 @@ import com.systematic.trading.model.price.TradingDayPrices;
 import com.systematic.trading.signals.data.api.alpha.vantage.converter.AlphaVantageResponseConverter;
 import com.systematic.trading.signals.data.api.alpha.vantage.dao.AlphaVantageApiDao;
 import com.systematic.trading.signals.data.api.alpha.vantage.dao.AlphaVantageApiFormatter;
-import com.systematic.trading.signals.data.api.alpha.vantage.resource.DigitalCurrencyDailyAudResource;
-import com.systematic.trading.signals.data.api.alpha.vantage.resource.ErrorPossibleResponseResource;
+import com.systematic.trading.signals.data.api.alpha.vantage.resource.AlphaVantageTimeSeriesDailyResponseResource;
+import com.systematic.trading.signals.data.api.alpha.vantage.resource.AlphaVantageTimeSeriesDigitalCurrencyDailyAudResource;
+import com.systematic.trading.signals.data.api.alpha.vantage.resource.AlphaVantageTimeSeriesEntries;
+import com.systematic.trading.signals.data.api.alpha.vantage.resource.AlphaVantqgeErrorPossibleResponseResource;
 
 /**
  * HttpAlphaVantageApiDao retrieves time series data from the Alpha Vantage API.
  * 
  * @author CJ Hare
  */
-public class HttpAlphaVantageApiDao implements AlphaVantageApiDao {
+public class HttpAlphaVantageApiDao<T> implements AlphaVantageApiDao {
 
 	/** Class' little logger, */
 	private static final Logger LOG = LogManager.getLogger(HttpAlphaVantageApiDao.class);
@@ -109,19 +111,19 @@ public class HttpAlphaVantageApiDao implements AlphaVantageApiDao {
 		final WebTarget url = apiFormatter.url(root, tickerDataset, tickerSymbol, apiKey);
 		final Response response = get(url, throttler);
 
-		final DigitalCurrencyDailyAudResource resource = response.readEntity(DigitalCurrencyDailyAudResource.class);
-		// TODO Or this could be
-		// final AlphaVantageTimeSeriesDailyResponseResource resource =
-		// response.readEntity(AlphaVantageTimeSeriesDailyResponseResource.class);
+		final AlphaVantageTimeSeriesEntries resource;
+		if ("AUD".equalsIgnoreCase(tickerDataset)) {
+			resource = response.readEntity(AlphaVantageTimeSeriesDigitalCurrencyDailyAudResource.class);
+		} else {
+			resource = response.readEntity(AlphaVantageTimeSeriesDailyResponseResource.class);
+		}
 
 		errorCheck(url, resource);
 
-		return converter.convert(tickerSymbol, startDateInclusive, endDateExclusive, resource.dataset());
-		//TODO need or this could be
-		// return converter.convert(tickerSymbol, startDateInclusive, endDateExclusive, resource.dataset());
+		return converter.convert(tickerSymbol, startDateInclusive, endDateExclusive, resource.data());
 	}
 
-	private void errorCheck( final WebTarget url, final ErrorPossibleResponseResource resource )
+	private void errorCheck( final WebTarget url, final AlphaVantqgeErrorPossibleResponseResource resource )
 	        throws CannotRetrieveDataException {
 
 		final Optional<String> error = resource.error();
